ollama_sdk_test/requirements.txt
# Ollama SDK Test Dependencies
ollama>=0.5.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
httpx>=0.24.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0